Moved suggest to lib
